refactor(server): the hot-update mechanism becomes @prismshadow/penguin-hmr
The version store, the atomic harness.json commit, the resource registry and the
park → boot → swap move into packages/hmr — a package that cannot see a platform.
HmrHost takes the packaged bundle as a constructor argument and is generic over
the api a platform exposes, so nothing in there can name a route, a service or a
plugin even by accident.

Its README carries the rule: mechanism only, and not to be changed without asking.
A package boundary enforces what discipline had been enforcing — and this is the
one package a broken hot update has to survive, since everything in it ships by
rebuilding and redeploying every installation.

What stays with the server is its half of the layer: the capability contract, the
upgrade endpoints, the seam, and the platform itself. The published
./hmr/manifest subpath stays too, as a re-export — installed CLIs resolve their
bundle through it.

Internal (private) and inlined into the server's bundle: a boundary in the source
tree, not another artifact on npm.
