+++ changelog/unreleased/2026-09-07-hmr-layer-and-package.zh.md
@@ -1,0 +1,17 @@
+# 这一层改名叫 HMR，它的机制独立成包
+
+- **Date:** 2026-09-07
+- **Type:** improvement
+- **Scope:** `server`, `desktop`
+
+[English](2026-09-07-hmr-layer-and-package.md)
+
+两处改动，同一个目的：让「把产品行为放进只能靠重装才能送达的地方」这件事变难。
+
+**这一层按它所做的事命名。** 「runtime」同时也指「正在运行的那个程序」，于是进程做的任何事听起来都像属于这一层——而落进这一层的东西，只能靠重建并重新分发每一个安装来发布。现在它在文档与标识符里都叫 HMR 层。资源 id 字符串仍保留 `runtime:` 前缀，这是刻意的：id 是跨代际的线上契约，改名会让旧层的注册对新平台不可见。
+
+**机制迁入 `packages/hmr`。** 版本仓库、`harness.json` 的原子提交、资源注册表，以及 park → boot → swap，现在住在一个**看不见 platform** 的包里：编译进程序的那份 bundle 是构造函数参数，platform 暴露的 api 是类型参数。它的 README 写明了规矩——只装机制，未经允许不得修改——而真正强制这条规矩的是包边界，不是自觉。
+
+留在 server 的是这一层属于产品的那一半：platform 能 claim 哪些能力、升级端点、HTTP 接缝，以及 platform 本身。
+
+它是仓库内部包，并被内联进 server 的产物：这是源码树里的边界，不是又一个要发布到 npm 的包。
