+++ packages/hmr/src/resources.ts
