+++ packages/server/package.json
@@ -72,6 +72,7 @@
     "@hono/node-server": "^2.0.5",
     "@larksuiteoapi/node-sdk": "^1.73.0",
     "@prismshadow/penguin-core": "workspace:*",
+    "@prismshadow/penguin-hmr": "workspace:*",
     "@xterm/headless": "^5.5.0",
     "dotenv": "^17.0.0",
     "fflate": "^0.8.3",
