+++ packages/server/src/api/types.ts
@@ -25,7 +25,7 @@ import type {
 } from "@prismshadow/penguin-core/interfaces";
 // Build/harness identity is not an interface contract — it ships from the barrel (core's version-info.ts).
 import type { HarnessInfo, VersionReport, HarnessHistory } from "@prismshadow/penguin-core";
-import type { IfacesDiff } from "../hmr/ifaces-diff.js";
+import type { IfacesDiff } from "@prismshadow/penguin-hmr";
 import type { WorkflowInfo } from "../mechanisms/workflows.js";
 import type { SandboxSettings as SandboxSettingsType } from "@prismshadow/penguin-core/plugin";
 import type {
@@ -3350,7 +3350,7 @@ export interface VersionRollbackResponse {
 
 /** GET /api/version/history/diff?from=&to=: what changed between two stored interface tables. */
 export type VersionHistoryDiffResponse = IfacesDiff;
-export type { IfaceChange, IfacesDiff, MemberChange, ModuleChange } from "../hmr/ifaces-diff.js";
+export type { IfaceChange, IfacesDiff, MemberChange, ModuleChange } from "@prismshadow/penguin-hmr";
 
 /**
  * GET /api/version/update-check: newest published release vs the running version.
