+++ packages/server/src/terminal/ws.ts
@@ -16,7 +16,7 @@ import type { IncomingMessage, Server as HttpServer } from "node:http";
 import type { Duplex } from "node:stream";
 import { WebSocketServer } from "ws";
 import { SESSION_COOKIE } from "../auth/middleware.js";
-import type { HmrHost } from "../hmr/host.js";
+import type { ServerHmrHost } from "../hmr/platform.js";
 import type { Auth } from "../mechanisms/identity.js";
 
 const STREAM_PATH = /^\/api\/terminals\/([^/]+)\/stream$/;
@@ -38,7 +38,7 @@ const BACKPRESSURE_POLL_MS = 250;
 
 export interface TerminalWebSocketDeps {
   /** The booted platform owns the terminals; the runtime asks it per upgrade. */
-  hmr: HmrHost;
+  hmr: ServerHmrHost;
   authService: Auth;
   log: (line: string) => void;
 }
